modify pom to add fat-jar build, integrate IDE run
modify pom to build fat jar and include resources . and main/resources
copy webroot to . to be included in fat-jar so that running java -jar
classpath will pick up webroot.

Modify runner.java to start at .; this will java app running in eclipse to
resolve the path to webroot and not result in 'not found' in web pages.

update pom

update readme